src/main/java/com/druidkuma/leetcode/spotify/RemoveAllAdjacentDuplicatesInStringII.java
package com.druidkuma.leetcode.spotify;

import java.util.ArrayDeque;
import java.util.Deque;

/**
 * 1209
 *
 * You are given a string s and an integer k, a k duplicate removal consists of choosing k adjacent and equal letters from s and removing them, causing the left and the right side of the deleted substring to concatenate together.
 *
 * We repeatedly make k duplicate removals on s until we no longer can.
 *
 * Return the final string after all such duplicate removals have been made. It is guaranteed that the answer is unique.
 */
public class RemoveAllAdjacentDuplicatesInStringII {
    public String removeDuplicates(String s, int k) {
        StringBuilder sb = new StringBuilder(s);
        Deque<Integer> counts = new ArrayDeque<>();
        for (int i = 0; i < sb.length(); i++) {
            if (i == 0 || sb.charAt(i) != sb.charAt(i - 1)) {
                counts.push(1);
            } else {
                int incremented = counts.pop() + 1;
                if (incremented == k) {
                    sb.delete(i - k + 1, i + 1);
                    i = i - k;
                } else {
                    counts.push(incremented);
                }
            }
        }
        return sb.toString();
    }

    public String removeDuplicatesBruteForce(String s, int k) {
        if (s.length() < k) return s;
        int count = 1;
        for (int i = 1; i < s.length(); i++) {
            if (s.charAt(i) == s.charAt(i-1)) count++;
            else {
                count = 1;
            }
            if (count == k) return removeDuplicates(s.substring(0, i-k+1) + s.substring(i + 1), k);
        }
        return s;
    }
}
